+++ internal/http/validation/init.go
@@ -1,4 +1,4 @@
-package validators
+package validation
 
 import (
 	"github.com/go-playground/validator/v10"
