+++ internal/http/validation/password_validator.go
@@ -1,4 +1,4 @@
-package validators
+package validation
 
 import (
 	"unicode"
