+++ internal/http/validation/translate.go
@@ -1,4 +1,4 @@
-package validators
+package validation
 
 import (
 	"errors"
